Move intent registration to own integration.
