+++ homeassistant/components/shopping_list/__init__.py
@@ -9,7 +9,6 @@ from homeassistant.const import HTTP_NOT_FOUND, HTTP_BAD_REQUEST
 from homeassistant.core import callback
 from homeassistant.components import http
 from homeassistant.components.http.data_validator import RequestDataValidator
-from homeassistant.helpers import intent
 import homeassistant.helpers.config_validation as cv
 from homeassistant.util.json import load_json, save_json
 from homeassistant.components import websocket_api
@@ -20,8 +19,6 @@ DOMAIN = "shopping_list"
 _LOGGER = logging.getLogger(__name__)
 CONFIG_SCHEMA = vol.Schema({DOMAIN: {}}, extra=vol.ALLOW_EXTRA)
 EVENT = "shopping_list_updated"
-INTENT_ADD_ITEM = "HassShoppingListAddItem"
-INTENT_LAST_ITEMS = "HassShoppingListLastItems"
 ITEM_UPDATE_SCHEMA = vol.Schema({"complete": bool, ATTR_NAME: str})
 PERSISTENCE = ".shopping_list.json"
 
@@ -85,9 +82,6 @@ def async_setup(hass, config):
 
     data = hass.data[DOMAIN] = ShoppingData(hass)
     yield from data.async_load()
-
-    intent.async_register(hass, AddItemIntent())
-    intent.async_register(hass, ListTopItemsIntent())
 
     hass.services.async_register(
         DOMAIN, SERVICE_ADD_ITEM, add_item_service, schema=SERVICE_ITEM_SCHEMA
@@ -173,49 +167,6 @@ class ShoppingData:
     def save(self):
         """Save the items."""
         save_json(self.hass.config.path(PERSISTENCE), self.items)
-
-
-class AddItemIntent(intent.IntentHandler):
-    """Handle AddItem intents."""
-
-    intent_type = INTENT_ADD_ITEM
-    slot_schema = {"item": cv.string}
-
-    @asyncio.coroutine
-    def async_handle(self, intent_obj):
-        """Handle the intent."""
-        slots = self.async_validate_slots(intent_obj.slots)
-        item = slots["item"]["value"]
-        intent_obj.hass.data[DOMAIN].async_add(item)
-
-        response = intent_obj.create_response()
-        response.async_set_speech(f"I've added {item} to your shopping list")
-        intent_obj.hass.bus.async_fire(EVENT)
-        return response
-
-
-class ListTopItemsIntent(intent.IntentHandler):
-    """Handle AddItem intents."""
-
-    intent_type = INTENT_LAST_ITEMS
-    slot_schema = {"item": cv.string}
-
-    @asyncio.coroutine
-    def async_handle(self, intent_obj):
-        """Handle the intent."""
-        items = intent_obj.hass.data[DOMAIN].items[-5:]
-        response = intent_obj.create_response()
-
-        if not items:
-            response.async_set_speech("There are no items on your shopping list")
-        else:
-            response.async_set_speech(
-                "These are the top {} items on your shopping list: {}".format(
-                    min(len(items), 5),
-                    ", ".join(itm["name"] for itm in reversed(items)),
-                )
-            )
-        return response
 
 
 class ShoppingListView(http.HomeAssistantView):
