+++ tests/components/shopping_list/test_init.py
@@ -1,27 +1,13 @@
 """Test shopping list component."""
 import asyncio
-from unittest.mock import patch
 
-import pytest
-
-from homeassistant.bootstrap import async_setup_component
 from homeassistant.helpers import intent
 from homeassistant.components.websocket_api.const import TYPE_RESULT
 
 
-@pytest.fixture(autouse=True)
-def mock_shopping_list_io():
-    """Stub out the persistence."""
-    with patch("homeassistant.components.shopping_list.ShoppingData.save"), patch(
-        "homeassistant.components.shopping_list." "ShoppingData.async_load"
-    ):
-        yield
-
-
 @asyncio.coroutine
-def test_add_item(hass):
+def test_add_item(hass, sl_setup):
     """Test adding an item intent."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     response = yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -31,9 +17,8 @@ def test_add_item(hass):
 
 
 @asyncio.coroutine
-def test_recent_items_intent(hass):
+def test_recent_items_intent(hass, sl_setup):
     """Test recent items."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -54,9 +39,8 @@ def test_recent_items_intent(hass):
 
 
 @asyncio.coroutine
-def test_deprecated_api_get_all(hass, hass_client):
+def test_deprecated_api_get_all(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -77,9 +61,8 @@ def test_deprecated_api_get_all(hass, hass_client):
     assert not data[1]["complete"]
 
 
-async def test_ws_get_items(hass, hass_ws_client):
+async def test_ws_get_items(hass, hass_ws_client, sl_setup):
     """Test get shopping_list items websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
 
     await intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -106,9 +89,8 @@ async def test_ws_get_items(hass, hass_ws_client):
 
 
 @asyncio.coroutine
-def test_deprecated_api_update(hass, hass_client):
+def test_deprecated_api_update(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -142,9 +124,8 @@ def test_deprecated_api_update(hass, hass_client):
     assert wine == {"id": wine_id, "name": "wine", "complete": True}
 
 
-async def test_ws_update_item(hass, hass_ws_client):
+async def test_ws_update_item(hass, hass_ws_client, sl_setup):
     """Test update shopping_list item websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
     await intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
     )
@@ -186,9 +167,8 @@ async def test_ws_update_item(hass, hass_ws_client):
 
 
 @asyncio.coroutine
-def test_api_update_fails(hass, hass_client):
+def test_api_update_fails(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -209,9 +189,8 @@ def test_api_update_fails(hass, hass_client):
     assert resp.status == 400
 
 
-async def test_ws_update_item_fail(hass, hass_ws_client):
+async def test_ws_update_item_fail(hass, hass_ws_client, sl_setup):
     """Test failure of update shopping_list item websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
     await intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
     )
@@ -234,9 +213,8 @@ async def test_ws_update_item_fail(hass, hass_ws_client):
 
 
 @asyncio.coroutine
-def test_deprecated_api_clear_completed(hass, hass_client):
+def test_deprecated_api_clear_completed(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     yield from intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
@@ -265,9 +243,8 @@ def test_deprecated_api_clear_completed(hass, hass_client):
     assert items[0] == {"id": wine_id, "name": "wine", "complete": False}
 
 
-async def test_ws_clear_items(hass, hass_ws_client):
+async def test_ws_clear_items(hass, hass_ws_client, sl_setup):
     """Test clearing shopping_list items websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
     await intent.async_handle(
         hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
     )
@@ -296,9 +273,8 @@ async def test_ws_clear_items(hass, hass_ws_client):
 
 
 @asyncio.coroutine
-def test_deprecated_api_create(hass, hass_client):
+def test_deprecated_api_create(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     client = yield from hass_client()
     resp = yield from client.post("/api/shopping_list/item", json={"name": "soda"})
@@ -315,9 +291,8 @@ def test_deprecated_api_create(hass, hass_client):
 
 
 @asyncio.coroutine
-def test_deprecated_api_create_fail(hass, hass_client):
+def test_deprecated_api_create_fail(hass, hass_client, sl_setup):
     """Test the API."""
-    yield from async_setup_component(hass, "shopping_list", {})
 
     client = yield from hass_client()
     resp = yield from client.post("/api/shopping_list/item", json={"name": 1234})
@@ -326,9 +301,8 @@ def test_deprecated_api_create_fail(hass, hass_client):
     assert len(hass.data["shopping_list"].items) == 0
 
 
-async def test_ws_add_item(hass, hass_ws_client):
+async def test_ws_add_item(hass, hass_ws_client, sl_setup):
     """Test adding shopping_list item websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
     client = await hass_ws_client(hass)
     await client.send_json({"id": 5, "type": "shopping_list/items/add", "name": "soda"})
     msg = await client.receive_json()
@@ -342,9 +316,8 @@ async def test_ws_add_item(hass, hass_ws_client):
     assert items[0]["complete"] is False
 
 
-async def test_ws_add_item_fail(hass, hass_ws_client):
+async def test_ws_add_item_fail(hass, hass_ws_client, sl_setup):
     """Test adding shopping_list item failure websocket command."""
-    await async_setup_component(hass, "shopping_list", {})
     client = await hass_ws_client(hass)
     await client.send_json({"id": 5, "type": "shopping_list/items/add", "name": 123})
     msg = await client.receive_json()
