+++ tests/components/shopping_list/test_intent.py
@@ -1,0 +1,22 @@
+"""Test Shopping List intents."""
+from homeassistant.helpers import intent
+
+
+async def test_recent_items_intent(hass, sl_setup):
+    """Test recent items."""
+    await intent.async_handle(
+        hass, "test", "HassShoppingListAddItem", {"item": {"value": "beer"}}
+    )
+    await intent.async_handle(
+        hass, "test", "HassShoppingListAddItem", {"item": {"value": "wine"}}
+    )
+    await intent.async_handle(
+        hass, "test", "HassShoppingListAddItem", {"item": {"value": "soda"}}
+    )
+
+    response = await intent.async_handle(hass, "test", "HassShoppingListLastItems")
+
+    assert (
+        response.speech["plain"]["speech"]
+        == "These are the top 3 items on your shopping list: soda, wine, beer"
+    )
